requirements.txt
streamlit>=1.35.0
google-generativeai>=0.8.0
pandas>=2.0.0
numpy>=1.26.0,<2.0.0
Pillow>=10.0.0
pdfplumber>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.10
opencv-python-headless>=4.9.0
requests>=2.28.0
gtts>=2.3.0
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99